Split up sessions and sessions-worker
